[CIR] Replace the complex lowering pass with a pipeline reorder
The dedicated pass expands complex mul and div ahead of the classifier, which
fixes the helper calls and nothing else.  Moving LoweringPrepare above
CallConvLowering does the same for every runtime call and every outlined global
initializer it creates.  The function addresses it takes all go through
`cir.get_global`, which the classifier already retypes for a rewritten callee.

LoweringPrepare, Passes.td, Passes.h, CIRTransformUtils and the transforms
CMakeLists go back to the state they had before this branch, so the pass
ordering in CIRPasses.cpp is the only functional change left.
