+++ clang/include/clang/CIR/Dialect/Passes.h
@@ -42,8 +42,6 @@ std::unique_ptr<Pass> createCallConvLoweringPass(
 std::unique_ptr<Pass> createHoistAllocasPass();
 std::unique_ptr<Pass> createLoweringPreparePass();
 std::unique_ptr<Pass> createLoweringPreparePass(clang::ASTContext *astCtx);
-std::unique_ptr<Pass> createComplexLoweringPass();
-std::unique_ptr<Pass> createComplexLoweringPass(clang::ASTContext *astCtx);
 std::unique_ptr<Pass> createGotoSolverPass();
 std::unique_ptr<Pass> createIdiomRecognizerPass();
 std::unique_ptr<Pass> createLibOptPass();
