+++ clang/test/CIR/CodeGen/complex-libcall-abi-global-init.cpp
@@ -10,9 +10,7 @@ extern float _Complex b;
 
 float _Complex g = a / b;
 
-// The coercion slot is allocated in the global initializer function.
-
-// CIR-LABEL: cir.func {{.*}}@__cxx_global_var_init
+// CIR-LABEL: cir.func {{.*}}@__cxx_global_var_init()
 // CIR: %[[SLOT:.*]] = cir.alloca "coerce" align(8) : !cir.ptr<!cir.vector<2 x !cir.float>>
 // CIR: %[[G:.*]] = cir.get_global @g : !cir.ptr<!cir.complex<!cir.float>>
 // CIR: %[[COERCED:.*]] = cir.call @__divsc3({{.*}}) : (!cir.float, !cir.float, !cir.float, !cir.float) -> !cir.vector<2 x !cir.float>
@@ -28,16 +26,21 @@ float _Complex g = a / b;
 // LLVMCIR: %[[RESULT:.+]] = load { float, float }, ptr %[[SLOT]], align 4
 // LLVMCIR: store { float, float } %[[RESULT]], ptr @g, align 4
 
-// OGCG: %[[COERCED:.*]] = call noundef <2 x float> @__divsc3(float noundef %{{.+}}, float noundef %{{.+}}, float noundef %{{.+}}, float noundef %{{.+}})
+// OGCG: call noundef <2 x float> @__divsc3(float noundef %{{.+}}, float noundef %{{.+}}, float noundef %{{.+}}, float noundef %{{.+}})
 
 float _Complex h = a * b;
 
 // Multiply puts its call inside the NaN-check region, one level further in, so
 // the slot has to be hoisted to the initializer function's entry block.
 
+// CIR-LABEL: cir.func {{.*}}@__cxx_global_var_init.1()
+// CIR: %[[SLOT:.*]] = cir.alloca "coerce" align(8) : !cir.ptr<!cir.vector<2 x !cir.float>>
+// CIR: %[[COERCED:.*]] = cir.call @__mulsc3({{.*}}) : (!cir.float, !cir.float, !cir.float, !cir.float) -> !cir.vector<2 x !cir.float>
+// CIR: cir.store %[[COERCED]], %[[SLOT]] : !cir.vector<2 x !cir.float>, !cir.ptr<!cir.vector<2 x !cir.float>>
+
 // LLVM-LABEL: define internal void @__cxx_global_var_init.1()
 // LLVMCIR: %[[SLOT:.+]] = alloca <2 x float>, align 8
-// LLVMCIR: br i1 %{{.+}}, label %[[THEN:.+]], label %[[ELSE:.+]]
+// LLVMCIR: br i1 %{{.+}}, label %[[THEN:.+]], label %{{.+}}
 // LLVMCIR: [[THEN]]:
 // LLVMCIR: %[[COERCED:.*]] = call <2 x float> @__mulsc3(float %{{.+}}, float %{{.+}}, float %{{.+}}, float %{{.+}})
 // LLVMCIR: store <2 x float> %[[COERCED]], ptr %[[SLOT]], align 8
